requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
asgiref==3.3.4
astroid==2.5.6
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
convertdate==2.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.7.0
django-ip-geolocation==1.6.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
hijri-converter==2.1.1
holidays==0.11.1
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
mysql-connector-python==8.0.24
mysqlclient==2.0.3
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.18
protobuf==3.15.8
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
PyJWT==1.7.1
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
PyMeeus==0.5.11
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3
